+++ backend/src/modules/experience/router.ts
@@ -1,9 +1,7 @@
 import express from 'express';
 import type { EntitySchema } from 'typeorm';
 import dataSource from '../../dataSource';
-import { createController } from '../common/controller/controller.factory';
-import { createRepository } from '../common/repository/repository.factory';
-import { createService } from '../common/service/service.factory';
+import { createController, createRepository, createService } from '../common';
 import { Experience } from './entity';
 
 const dataRepository = createRepository<Experience>(
